Refactor the code and move functions to utils file
